chore: restructure files/folders
