+++ ts/logic/utils.spec.ts
@@ -1,4 +1,4 @@
-import { assertNonNullish } from "./assertNonNullish";
+import { assertNonNullish, roundToTwoDecimals } from "./utils";
 
 describe("assertNonNullish", () => {
   it("throws an error when the value is null or undefined", () => {
@@ -14,5 +14,18 @@ describe("assertNonNullish", () => {
     assertNonNullish(1, "Value cannot be null");
     assertNonNullish("test", "Value cannot be undefined");
     assertNonNullish({}, "Value cannot be null or undefined");
+  });
+});
+
+describe("roundToTwoDecimals", () => {
+  describe("given a number", () => {
+    it("should return a number with two decimals", () => {
+      expect(roundToTwoDecimals(9)).toBe(9);
+      expect(roundToTwoDecimals(9.9111)).toBe(9.91);
+      expect(roundToTwoDecimals(9.911)).toBe(9.91);
+      expect(roundToTwoDecimals(9.91)).toBe(9.91);
+      expect(roundToTwoDecimals(0)).toBe(0);
+      expect(typeof roundToTwoDecimals(8.123)).toBe("number");
+    });
   });
 });
