+++ ts/logic/utils.ts
@@ -7,3 +7,7 @@ export function assertNonNullish<T>(
   }
   return;
 }
+
+export function roundToTwoDecimals(value: number) {
+  return Number(Math.round(parseFloat(value + "e" + 2)) + "e-" + 2);
+}
