--- ts/logic/utils/generalFunctions.ts
@@ -1,2 +1,0 @@
-export const roundToTwoDecimals = (value: number): number =>
-  Number(Math.round(parseFloat(value + "e" + 2)) + "e-" + 2);
